+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.0.5 (`2023-05-31`)
+- add: MetalLB as load balancer
+- add: don't use native Traefik Ingress
+
 # 0.0.4 (`2023-05-31`)
 - add: nfs-utils for k3s-cluster
 - add: k3s-nfs-provisoner systemd service